Create a singleton instance of the `OAuthHandler`
As pointed out in the comment, this solution is far from ideal.
But since we need to access the handler from the `SceneDelegate`
(for handling the URL), and since we do not have a proper
injection mechanism at the moment, this is a quick way
to allow for this access.